+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [8.18.0] - 2026-06-16
+
 ### Added
 
 - `gen circleci`: new `--image-name` and `--image-platforms` flags for repos whose image pipeline
@@ -2162,7 +2164,8 @@ Renovate config
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/devctl/compare/v8.17.1...HEAD
+[Unreleased]: https://github.com/giantswarm/devctl/compare/v8.18.0...HEAD
+[8.18.0]: https://github.com/giantswarm/devctl/compare/v8.17.1...v8.18.0
 [8.17.1]: https://github.com/giantswarm/devctl/compare/v8.17.0...v8.17.1
 [8.17.0]: https://github.com/giantswarm/devctl/compare/v8.16.0...v8.17.0
 [8.16.0]: https://github.com/giantswarm/devctl/compare/v8.15.2...v8.16.0